+++ ktlint/src/test/kotlin/com/pinterest/ktlint/RuleSetsLoaderCLITest.kt
@@ -36,12 +36,10 @@ class RuleSetsLoaderCLITest {
                 listOf("-R", "$tempDir/$jarWithRulesetProviderV2"),
             ) {
                 SoftAssertions().apply {
-                    assertNormalExitCode()
-                    // JAR ruleset provided with path "/var/folders/24/wtp_g21953x22nr8z86gvltc0000gp/T/junit920502858262478102/custom-ruleset/rule-set-provider-v2/ktlint-ruleset-template.jar
-                    // Add editor config override to enable rule set(s) '[indent-string-template-ruleset]' from custom rule set JAR('s): '[/var/folders/24/wtp_g21953x22nr8z86gvltc0000gp/T/junit920502858262478102/custom-ruleset/rule-set-provider-v2/ktlint-ruleset-template.jar]'
+                    assertErrorExitCode()
                     assertThat(normalOutput)
                         .containsLineMatching(Regex(".* JAR ruleset provided with path .*$jarWithRulesetProviderV2.*"))
-                        .containsLineMatching(Regex(".* Add editor config override to enable rule set\\(s\\) '\\[indent-string-template-ruleset]' from custom rule set JAR\\('s\\): .*$jarWithRulesetProviderV2.*"))
+                        .containsLineMatching(Regex(".*custom-ruleset.rule-set-provider-v2.Main.kt:1:1: Unexpected var, use val instead.*custom-rule-set-id:no-var.*"))
                 }.assertAll()
             }
     }