+++ ktlint/src/test/resources/cli/custom-ruleset/rule-set-provider-v2/Main.kt
@@ -1,3 +1,5 @@
+var helloWorld = "Hello world!"
+
 fun main() {
-    println("Hello world!")
+    println(helloWorld)
 }